LayoutTests/fast/js/native-error-prototype-expected.txt

This is a test case for bug 55346.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS ({}).toString.call(Error.prototype) is "[object Error]"
PASS ({}).toString.call(RangeError.prototype) is "[object Error]"
PASS successfullyParsed is true

TEST COMPLETE
